requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.1

# LangChain and LLM - updated to recent compatible versions
langchain>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0

# Search tool
tavily-python>=0.3.0

# Data processing - updated for Python 3.13 compatibility
pandas>=2.2.0

# HTTP client for examples
requests>=2.32.0
aiohttp>=3.9.1 